Nutrient_Deficiency_and_Disease_Symptoms_in_Rice_Classification/deployment/requirements.txt
streamlit
tensorflow-cpu
numpy
pandas
Image